[FIX] security and refactor
