+++ src/index.js
@@ -12,8 +12,9 @@ import session from "express-session";
 import { corsOptions } from "./common/config/cors.config.js";
 import { MainRouter } from "./api/routes.js";
 import swaggerDocs from "./swagger.js";
-import { securityMiddleware } from "./common/middleware/security.middleware.js";
+import { securityMiddleware } from "./common/middlewares/security.middleware.js";
 import redisService from './common/services/redis.service.js';
+import { csrfProtection } from "./common/middlewares/csrf.middleware.js";
 
 // Create __dirname equivalent for ES modules
 const __filename = fileURLToPath(import.meta.url);
@@ -70,6 +71,8 @@ app.use(session({
     maxAge: 24 * 60 * 60 * 1000 // 24 hours
   }
 }));
+
+app.use(csrfProtection())
 
 app.use(morgan('dev', {
   skip: function (req, res) { return res.statusCode < 400 }
